MetaLookup interface setup #517
Allow local/remote implementations in the future
