+++ crnk-data/crnk-data-jpa/src/test/java/io/crnk/data/jpa/meta/MetaEntityTest.java
@@ -4,7 +4,7 @@ import io.crnk.data.jpa.model.SequenceEntity;
 import io.crnk.data.jpa.model.TestSubclassWithSuperclassGenerics;
 import io.crnk.data.jpa.model.TestSubclassWithSuperclassPk;
 import io.crnk.data.jpa.model.TestMappedSuperclassWithPk;
-import io.crnk.meta.MetaLookup;
+import io.crnk.meta.MetaLookupImpl;
 import io.crnk.meta.model.MetaAttribute;
 import io.crnk.meta.model.MetaDataObject;
 import io.crnk.meta.model.MetaPrimaryKey;
@@ -23,7 +23,7 @@ public class MetaEntityTest {
     @Before
     public void setup() {
         metaProvider = new JpaMetaProvider(Collections.emptySet());
-        MetaLookup lookup = new MetaLookup();
+        MetaLookupImpl lookup = new MetaLookupImpl();
         lookup.addProvider(metaProvider);
     }
