+++ crnk-meta/src/main/java/io/crnk/meta/MetaModule.java
@@ -47,225 +47,225 @@ import java.util.Set;
 
 public class MetaModule implements ModuleExtensionAware<MetaModuleExtension> {
 
-	/**
-	 * Current MetaLookup instance to be used.
-	 */
-	private MetaLookup currentLookup;
+    /**
+     * Current MetaLookup instance to be used.
+     */
+    private MetaLookupImpl currentLookup;
 
-	/**
-	 * MetaLookup used by the current request to ensure consistent responses even in case of concurrent
-	 * modifications.
-	 */
-	// TODO move to request scoping once async is in palce
-	private ThreadLocal<MetaLookup> lookupRequestLocal = new ThreadLocal<>();
+    /**
+     * MetaLookup used by the current request to ensure consistent responses even in case of concurrent
+     * modifications.
+     */
+    // TODO move to request scoping once async is in palce
+    private ThreadLocal<MetaLookup> lookupRequestLocal = new ThreadLocal<>();
 
-	private ModuleContext context;
+    private ModuleContext context;
 
-	private MetaModuleConfig config;
+    private MetaModuleConfig config;
 
-	private DefaultResourceInformationProvider informationBuilder;
-
-
-	/**
-	 * @deprecated use {@link #createClientModule()} or {@link #createServerModule(MetaModuleConfig)}
-	 */
-	// make protected for CDI in the future and remove deprecation
-	@Deprecated
-	public MetaModule() {
-		this(new MetaModuleConfig());
-	}
-
-	private MetaModule(MetaModuleConfig config) {
-		this.config = config;
-	}
+    private DefaultResourceInformationProvider informationBuilder;
 
 
-	/**
-	 * @deprecated use {@link #createClientModule()} or {@link #createServerModule(MetaModuleConfig)}
-	 */
-	@Deprecated
-	public static MetaModule create() {
-		return new MetaModule();
-	}
+    /**
+     * @deprecated use {@link #createClientModule()} or {@link #createServerModule(MetaModuleConfig)}
+     */
+    // make protected for CDI in the future and remove deprecation
+    @Deprecated
+    public MetaModule() {
+        this(new MetaModuleConfig());
+    }
 
-	public static MetaModule createClientModule() {
-		return new MetaModule(new MetaModuleConfig());
-	}
-
-	public static MetaModule createServerModule(MetaModuleConfig config) {
-		return new MetaModule(config);
-	}
+    private MetaModule(MetaModuleConfig config) {
+        this.config = config;
+    }
 
 
-	@Override
-	public String getModuleName() {
-		return "meta";
-	}
+    /**
+     * @deprecated use {@link #createClientModule()} or {@link #createServerModule(MetaModuleConfig)}
+     */
+    @Deprecated
+    public static MetaModule create() {
+        return new MetaModule();
+    }
 
-	/**
-	 * @deprecated make use of {@link MetaModuleConfig} and pass to this instance upon creation
-	 */
-	public void addMetaProvider(MetaProvider provider) {
-		config.addMetaProvider(provider);
-	}
+    public static MetaModule createClientModule() {
+        return new MetaModule(new MetaModuleConfig());
+    }
 
-	@Override
-	public void setupModule(ModuleContext context) {
-		this.context = context;
+    public static MetaModule createServerModule(MetaModuleConfig config) {
+        return new MetaModule(config);
+    }
 
-		informationBuilder = registerInformationBuilder(context.getPropertiesProvider());
 
-		if (context.isServer()) {
-			context.addFilter(new DocumentFilter() {
-				@Override
-				public Response filter(DocumentFilterContext filterRequestContext, DocumentFilterChain chain) {
-					try {
-						return chain.doFilter(filterRequestContext);
-					} finally {
-						lookupRequestLocal.remove();
-					}
-				}
-			});
-		} else {
-			context.addResourceLookup(new ResourceLookup() {
+    @Override
+    public String getModuleName() {
+        return "meta";
+    }
 
-				@SuppressWarnings("unchecked")
-				@Override
-				public Set<Class<?>> getResourceClasses() {
-					return (Set) collectMetaClasses();
-				}
-			});
-		}
-	}
+    /**
+     * @deprecated make use of {@link MetaModuleConfig} and pass to this instance upon creation
+     */
+    public void addMetaProvider(MetaProvider provider) {
+        config.addMetaProvider(provider);
+    }
 
-	private void initRefreshListener() {
-		ResourceRegistry resourceRegistry = context.getResourceRegistry();
-		resourceRegistry.addListener(new ResourceRegistryPartAdapter() {
-			@Override
-			public void onChanged(ResourceRegistryPartEvent event) {
-				currentLookup = null;
-			}
-		});
-	}
+    @Override
+    public void setupModule(ModuleContext context) {
+        this.context = context;
 
-	protected DefaultResourceInformationProvider registerInformationBuilder(PropertiesProvider propertiesProvider) {
-		InformationBuilder informationBuilder = new DefaultInformationBuilder(context.getTypeParser());
-		DefaultResourceInformationProvider informationProvider = new DefaultResourceInformationProvider(
-				propertiesProvider,
-				new OffsetLimitPagingBehavior(),
-				new DefaultResourceFieldInformationProvider(),
-				new JacksonResourceFieldInformationProvider());
-		informationProvider.init(new DefaultResourceInformationProviderContext(informationProvider, informationBuilder,
-				context.getTypeParser(), null) {
-			@Override
-			public ObjectMapper getObjectMapper() {
-				return context.getObjectMapper();
-			}
-		});
-		return informationProvider;
-	}
+        informationBuilder = registerInformationBuilder(context.getPropertiesProvider());
 
-	protected void registerRepositories(DefaultResourceInformationProvider informationBuilder,
-										Set<Class<? extends MetaElement>> metaClasses) {
+        if (context.isServer()) {
+            context.addFilter(new DocumentFilter() {
+                @Override
+                public Response filter(DocumentFilterContext filterRequestContext, DocumentFilterChain chain) {
+                    try {
+                        return chain.doFilter(filterRequestContext);
+                    } finally {
+                        lookupRequestLocal.remove();
+                    }
+                }
+            });
+        } else {
+            context.addResourceLookup(new ResourceLookup() {
 
-		Supplier<MetaLookup> lookupSupplier = new Supplier<MetaLookup>() {
-			@Override
-			public MetaLookup get() {
-				return getLookup();
-			}
-		};
+                @SuppressWarnings("unchecked")
+                @Override
+                public Set<Class<?>> getResourceClasses() {
+                    return (Set) collectMetaClasses();
+                }
+            });
+        }
+    }
 
-		for (Class<? extends MetaElement> metaClass : metaClasses) {
-			context.addRepository(new MetaResourceRepositoryImpl<>(lookupSupplier, metaClass));
+    private void initRefreshListener() {
+        ResourceRegistry resourceRegistry = context.getResourceRegistry();
+        resourceRegistry.addListener(new ResourceRegistryPartAdapter() {
+            @Override
+            public void onChanged(ResourceRegistryPartEvent event) {
+                currentLookup = null;
+            }
+        });
+    }
 
-			HashSet<Class<? extends MetaElement>> targetResourceClasses = new HashSet<>();
-			ResourceInformation information = informationBuilder.build(metaClass);
-			for (ResourceField relationshipField : information.getRelationshipFields()) {
-				if (!MetaElement.class.isAssignableFrom(relationshipField.getElementType())) {
-					throw new IllegalStateException("only MetaElement relations supported, got " + relationshipField);
-				}
-				targetResourceClasses.add((Class<? extends MetaElement>) relationshipField.getElementType());
-			}
-			for (Class<? extends MetaElement> targetResourceClass : targetResourceClasses) {
-				context.addRepository(new MetaRelationshipRepositoryImpl(lookupSupplier, metaClass, targetResourceClass));
-			}
-		}
-	}
+    protected DefaultResourceInformationProvider registerInformationBuilder(PropertiesProvider propertiesProvider) {
+        InformationBuilder informationBuilder = new DefaultInformationBuilder(context.getTypeParser());
+        DefaultResourceInformationProvider informationProvider = new DefaultResourceInformationProvider(
+                propertiesProvider,
+                new OffsetLimitPagingBehavior(),
+                new DefaultResourceFieldInformationProvider(),
+                new JacksonResourceFieldInformationProvider());
+        informationProvider.init(new DefaultResourceInformationProviderContext(informationProvider, informationBuilder,
+                context.getTypeParser(), null) {
+            @Override
+            public ObjectMapper getObjectMapper() {
+                return context.getObjectMapper();
+            }
+        });
+        return informationProvider;
+    }
 
-	protected Set<Class<? extends MetaElement>> collectMetaClasses() {
-		final Set<Class<? extends MetaElement>> metaClasses = new HashSet<>();
-		metaClasses.add(MetaArrayType.class);
-		metaClasses.add(MetaAttribute.class);
-		metaClasses.add(MetaCollectionType.class);
-		metaClasses.add(MetaDataObject.class);
-		metaClasses.add(MetaElement.class);
-		metaClasses.add(MetaEnumType.class);
-		metaClasses.add(MetaInterface.class);
-		metaClasses.add(MetaKey.class);
-		metaClasses.add(MetaListType.class);
-		metaClasses.add(MetaLiteral.class);
-		metaClasses.add(MetaMapType.class);
-		metaClasses.add(MetaPrimaryKey.class);
-		metaClasses.add(MetaPrimitiveType.class);
-		metaClasses.add(MetaSetType.class);
-		metaClasses.add(MetaType.class);
+    protected void registerRepositories(DefaultResourceInformationProvider informationBuilder,
+                                        Set<Class<? extends MetaElement>> metaClasses) {
 
-		collectMetaClasses(metaClasses, config.getProviders());
-		return metaClasses;
-	}
+        Supplier<MetaLookup> lookupSupplier = new Supplier<MetaLookup>() {
+            @Override
+            public MetaLookup get() {
+                return getLookup();
+            }
+        };
 
-	private void collectMetaClasses(Set<Class<? extends MetaElement>> metaClasses, Collection<MetaProvider> providers) {
-		for (MetaProvider provider : providers) {
-			metaClasses.addAll(provider.getMetaTypes());
-			collectMetaClasses(metaClasses, provider.getDependencies());
-		}
-	}
+        for (Class<? extends MetaElement> metaClass : metaClasses) {
+            context.addRepository(new MetaResourceRepositoryImpl<>(lookupSupplier, metaClass));
 
-	@Override
-	public void setExtensions(List<MetaModuleExtension> extensions) {
-		for (MetaModuleExtension extension : extensions) {
-			for (MetaProvider provider : extension.getProviders()) {
-				config.addMetaProvider(provider);
-			}
-		}
-	}
+            HashSet<Class<? extends MetaElement>> targetResourceClasses = new HashSet<>();
+            ResourceInformation information = informationBuilder.build(metaClass);
+            for (ResourceField relationshipField : information.getRelationshipFields()) {
+                if (!MetaElement.class.isAssignableFrom(relationshipField.getElementType())) {
+                    throw new IllegalStateException("only MetaElement relations supported, got " + relationshipField);
+                }
+                targetResourceClasses.add((Class<? extends MetaElement>) relationshipField.getElementType());
+            }
+            for (Class<? extends MetaElement> targetResourceClass : targetResourceClasses) {
+                context.addRepository(new MetaRelationshipRepositoryImpl(lookupSupplier, metaClass, targetResourceClass));
+            }
+        }
+    }
 
-	@Override
-	public void init() {
-		final Set<Class<? extends MetaElement>> metaClasses = collectMetaClasses();
-		if (context.isServer()) {
-			initRefreshListener();
-			registerRepositories(informationBuilder, metaClasses);
-		}
-	}
+    protected Set<Class<? extends MetaElement>> collectMetaClasses() {
+        final Set<Class<? extends MetaElement>> metaClasses = new HashSet<>();
+        metaClasses.add(MetaArrayType.class);
+        metaClasses.add(MetaAttribute.class);
+        metaClasses.add(MetaCollectionType.class);
+        metaClasses.add(MetaDataObject.class);
+        metaClasses.add(MetaElement.class);
+        metaClasses.add(MetaEnumType.class);
+        metaClasses.add(MetaInterface.class);
+        metaClasses.add(MetaKey.class);
+        metaClasses.add(MetaListType.class);
+        metaClasses.add(MetaLiteral.class);
+        metaClasses.add(MetaMapType.class);
+        metaClasses.add(MetaPrimaryKey.class);
+        metaClasses.add(MetaPrimitiveType.class);
+        metaClasses.add(MetaSetType.class);
+        metaClasses.add(MetaType.class);
 
-	public synchronized MetaLookup getLookup() {
-		if (currentLookup == null) {
-			initLookup();
-		}
+        collectMetaClasses(metaClasses, config.getProviders());
+        return metaClasses;
+    }
 
-		MetaLookup requestLookup = lookupRequestLocal.get();
-		if (requestLookup == null) {
-			requestLookup = currentLookup;
-			lookupRequestLocal.set(requestLookup);
-		}
-		return requestLookup;
-	}
+    private void collectMetaClasses(Set<Class<? extends MetaElement>> metaClasses, Collection<MetaProvider> providers) {
+        for (MetaProvider provider : providers) {
+            metaClasses.addAll(provider.getMetaTypes());
+            collectMetaClasses(metaClasses, provider.getDependencies());
+        }
+    }
 
-	private void initLookup() {
-		MetaLookup lookup = new MetaLookup();
-		config.apply(lookup);
-		lookup.setModuleContext(context);
-		lookup.initialize();
-		currentLookup = lookup;
-	}
+    @Override
+    public void setExtensions(List<MetaModuleExtension> extensions) {
+        for (MetaModuleExtension extension : extensions) {
+            for (MetaProvider provider : extension.getProviders()) {
+                config.addMetaProvider(provider);
+            }
+        }
+    }
 
-	protected ThreadLocal<MetaLookup> getLookupRequestLocal() {
-		return lookupRequestLocal;
-	}
+    @Override
+    public void init() {
+        final Set<Class<? extends MetaElement>> metaClasses = collectMetaClasses();
+        if (context.isServer()) {
+            initRefreshListener();
+            registerRepositories(informationBuilder, metaClasses);
+        }
+    }
 
-	protected void reset() {
-		currentLookup = null;
-	}
+    public synchronized MetaLookup getLookup() {
+        if (currentLookup == null) {
+            initLookup();
+        }
+
+        MetaLookupImpl requestLookup = (MetaLookupImpl) lookupRequestLocal.get();
+        if (requestLookup == null) {
+            requestLookup = currentLookup;
+            lookupRequestLocal.set(requestLookup);
+        }
+        return requestLookup;
+    }
+
+    private void initLookup() {
+        MetaLookupImpl lookup = new MetaLookupImpl();
+        config.apply(lookup);
+        lookup.setModuleContext(context);
+        lookup.initialize();
+        currentLookup = lookup;
+    }
+
+    protected ThreadLocal<MetaLookup> getLookupRequestLocal() {
+        return lookupRequestLocal;
+    }
+
+    protected void reset() {
+        currentLookup = null;
+    }
 
 }
