+++ crnk-meta/src/main/java/io/crnk/meta/model/resource/MetaResource.java
@@ -8,13 +8,23 @@ import io.crnk.core.resource.annotations.JsonApiResource;
 @JsonApiResource(type = "meta/resource")
 public class MetaResource extends MetaResourceBase { // NOSONAR ignore exception hierarchy
 
-	private String resourceType;
+    private String resourceType;
 
-	public String getResourceType() {
-		return resourceType;
-	}
+    private String resourcePath;
 
-	public void setResourceType(String resourceType) {
-		this.resourceType = resourceType;
-	}
+    public String getResourceType() {
+        return resourceType;
+    }
+
+    public void setResourceType(String resourceType) {
+        this.resourceType = resourceType;
+    }
+
+    public String getResourcePath() {
+        return resourcePath;
+    }
+
+    public void setResourcePath(String resourcePath) {
+        this.resourcePath = resourcePath;
+    }
 }
