+++ crnk-operations/src/test/java/io/crnk/operations/AbstractOperationsTest.java
@@ -21,7 +21,7 @@ import io.crnk.data.jpa.meta.JpaMetaProvider;
 import io.crnk.data.jpa.query.JpaQueryFactory;
 import io.crnk.data.jpa.query.JpaQueryFactoryContext;
 import io.crnk.data.jpa.query.criteria.JpaCriteriaQueryFactory;
-import io.crnk.meta.MetaLookup;
+import io.crnk.meta.MetaLookupImpl;
 import io.crnk.meta.MetaModule;
 import io.crnk.meta.provider.MetaPartition;
 import io.crnk.meta.provider.resource.ResourceMetaProvider;
@@ -68,7 +68,7 @@ public abstract class AbstractOperationsTest extends JerseyTestBase {
 		factory.initalize(new JpaQueryFactoryContext() {
 			@Override
 			public MetaPartition getMetaPartition() {
-				MetaLookup metaLookup = new MetaLookup();
+				MetaLookupImpl metaLookup = new MetaLookupImpl();
 				JpaMetaProvider metaProvider = new JpaMetaProvider(em.getEntityManagerFactory());
 				metaLookup.addProvider(metaProvider);
 				metaLookup.initialize();
